Wifi: Allow to define the idle timeout (5 mn to 3 hours)
The setting key already exists in frameworks and is set to 15mn

This is only used when sleep policy allow to turn off the device.
Actually after 15mn, the driver is unloaded, and related services
fully stopped or paused.

Note: using the sleep policy can cause compatibility problems
with devices still using the old wpa supplicant (v6 or less),
In v8, ICS uses a new INTERFACE_DISABLED state for the wpa service.

Change-Id: I11953aab6e95bc3a2d5eaf1537acdf8a27701701
Signed-off-by: Tanguy Pruvot <tanguy.pruvot@gmail.com>

Conflicts:

	res/values/arrays.xml
	res/values/strings.xml